增加 v1/activity/name/{name}/with_coupon v1/coupon/by/category/{cid} API 接口

+++ src/main/java/com/lin/missyou/model/BannerItem.java
@@ -22,7 +22,7 @@ public class BannerItem extends Base {
     private Long id;
     private String img;
     private String keyword;
-    private short type;
+    private Integer type;
     private Long bannerId;
     private String name;
 }
